refactor: move signing config from signing/v1alpha1 to spec/signing/v1alpha1
Aligns with the pattern of spec/credentials/v1alpha1 and spec/identity/v1alpha1.
Update all import paths accordingly.

On-behalf-of: @SAP <jakob.moeller@sap.com>
On-behalf-of: @SAP <jakob.moeller@sap.com>
