+++ bindings/go/gpg/spec/signing/v1alpha1/algorithm.go
