+++ bindings/go/gpg/spec/signing/v1alpha1/config.go
