+++ bindings/go/gpg/spec/signing/v1alpha1/group_version.go
