+++ bindings/go/gpg/spec/signing/v1alpha1/zz_generated.ocm_type.go
